agents/pb-foundry-bing-agent/src/foundry_bing_agent/runtime/openai_client.py
from __future__ import annotations

from azure.ai.projects import AIProjectClient
from azure.identity import DefaultAzureCredential

from foundry_bing_agent.config import Settings


def build_project_client(settings: Settings) -> AIProjectClient:
    credential = DefaultAzureCredential()
    return AIProjectClient(endpoint=settings.azure_projects_endpoint, credential=credential)
