agents/pb-foundry-bing-agent/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundry-bing-agent"
version = "0.1.0"
description = "Foundry-only Bing-enabled agent runtime (no M365 channel)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "azure-ai-projects>=1.0.0b10",
  "azure-identity>=1.17.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.1",
  "pytest>=8.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
